fix: compile error

+++ src/include/tlct/convert/helper/functional.cpp
@@ -50,9 +50,9 @@ namespace rgs = std::ranges;
 
     const float sqrX = gradX * gradX;
     const float sqrY = gradY * gradY;
-    const float gradDeg30 = std::sqrtf(sqrX * 0.75f + sqrY * 0.25f);
-    const float gradDeg60 = std::sqrtf(sqrX * 0.25f + sqrY * 0.75f);
-    const float gradNormed = std::sqrtf(sqrX * 0.5f + sqrY * 0.5f);
+    const float gradDeg30 = std::sqrt(sqrX * 0.75f + sqrY * 0.25f);
+    const float gradDeg60 = std::sqrt(sqrX * 0.25f + sqrY * 0.75f);
+    const float gradNormed = std::sqrt(sqrX * 0.5f + sqrY * 0.5f);
 
     return {gradX, gradDeg30, gradDeg60, sqrY, gradNormed};
 }
